feat(entity-ai): 对齐MC 1.16.5实体AI系统并优化性能
## 核心对齐修复

### MobEntity旋转速度
- getHorizontalFaceSpeed(): 10.0f -> 75.0f
- getVerticalFaceSpeed(): 10.0f -> 40.0f
- getFaceRotSpeed(): 保持10.0f

### LookController
- 添加approachTargetAngle()函数用于导航路径限制
- 区分clampedRotate(用于目标跟踪)和approachTargetAngle(用于导航限制)
- 俯仰角重置为0.0f(原版行为)

### MovementController
- 修复setMoveTo()检查JUMPING状态
- 修复MOVE_TO旋转限制从30度到90度
- 添加STRAFE向量变换基于实体偏航角

### PathPoint
- 默认nodeType改为Blocked(原版行为)

### Sensor
- 使用mc::math::Random替代std::mt19937
- 添加initCounter()方法使用实体随机数初始化
- 修复counter初始化使用-1作为未初始化哨兵值

### GoalConstants
- 添加MC 1.16.5常量: VIEW_CHANGE_THRESHOLD, MOVEMENT_THRESHOLD
- 添加路径重算常量: PATH_RECALCULATE_BASE, PATH_RECALCULATE_RANDOM等
- 添加距离阈值: DISTANCE_FAR_THRESHOLD, DISTANCE_MEDIUM_THRESHOLD

### MeleeAttackGoal
- 修复攻击距离公式: (width * 2)^2 + targetWidth
- 添加路径重算逻辑with距离间隔
- 添加m_targetX/Y/Z, m_failedPathFindingPenalty, m_canPenalize

## 性能优化

- Brain::tickMemories(): 使用原地删除替代临时vector
- Brain::tickTasks()/stopAllTasks(): 直接遍历替代临时vector
- AvoidEntityGoal::findEscapePosition(): 将距离计算移出循环
- MeleeAttackGoal::attackTarget(): 移除重复攻击冷却设置
- TemptGoal: 使用常量替代硬编码魔法值

+++ src/common/entity/ai/goal/goals/TemptGoal.hpp
@@ -73,11 +73,11 @@ protected:
     ItemPredicate m_itemPredicate;
     bool m_scaredByMovement;
     Player* m_temptingPlayer = nullptr;
-    f32 m_targetX = 0.0f;
-    f32 m_targetY = 0.0f;
-    f32 m_targetZ = 0.0f;
-    f32 m_prevPitch = 0.0f;
-    f32 m_prevYaw = 0.0f;
+    f64 m_targetX = 0.0;
+    f64 m_targetY = 0.0;
+    f64 m_targetZ = 0.0;
+    f64 m_prevPitch = 0.0;
+    f64 m_prevYaw = 0.0;
     i32 m_delayTemptCounter = 0;
     bool m_isRunning = false;
